third_party/blink/web_tests/platform/win/external/wpt/css/css-fonts/generic-family-keywords-002-expected.txt

This is a testharness.js-based test.
[FAIL] font-family: -webkit-body treated as <font-family>, not <generic-name>
  assert_equals: expected 50 but got 25
Harness: the test ran to completion.
